refactor(creation): components folder
